LayoutTests/inspector/styles/paste-property-expected.txt

Tests that splitting properties when pasting works.

Text
Before pasting:
[expanded] 
element.style  { ()
font-size: 12px;

======== Matched CSS Rules ========
[expanded] 
div  { (user agent stylesheet)
display: block;


After pasting 'margin-left: 1px':
[expanded] 
element.style  { ()
margin-left: 1px;
font-size: 12px;

======== Matched CSS Rules ========
[expanded] 
div  { (user agent stylesheet)
display: block;


After pasting 'margin-top: 1px; color: red;':
[expanded] 
element.style  { ()
margin-left: 1px;
font-size: 12px;
margin-top: 1px;
color: red;

======== Matched CSS Rules ========
[expanded] 
div  { (user agent stylesheet)
display: block;


After pasting 'foo: bar; moo: zoo' over 'margin-top':
[expanded] 
element.style  { ()
margin-left: 1px;
font-size: 12px;
/-- overloaded --/ foo: bar;
/-- overloaded --/ moo: zoo;
color: red;

======== Matched CSS Rules ========
[expanded] 
div  { (user agent stylesheet)
display: block;
